Conflicts:
	GameProject/Base/Source/GameModel2D.cpp
	GameProject/Base/Source/GameView2D.cpp

+++ GameProject/Base/Source/Weapon.cpp
@@ -39,7 +39,7 @@ float CWeapon::GetCooldown(void)
 {
 	return FireCooldown;
 }
-float CWeapon::GetDamage(void)
+int CWeapon::GetDamage(void)
 {
 	return Damage;
 }
